[DIRTY HACK] Fix 'git' tracking on (Dockerized) Heroku dyno!
Unlike on a real machine, Dockerized deploys on Heroku do not add 'git' tracking information into it, which brings us an annoying bug where the updater falsely said it was up to date!
This patch fixes that by looking for the '.git' directory in the container's work directory, patching it with a hack if needed.